Refactor for use API v2 (#3)
* ajustes para a api v2

* .NET 9 Update

* Open Service Order

* Insert and Update Service Order

* fix app

+++ src/Fluxus.Core/Models/Invoice.cs
@@ -5,15 +5,15 @@ namespace Fluxus.Core.Models;
 
 public class Invoice
 {
-    public int Id { get; private set; }
+    public long Id { get; private set; }
     public string? Description { get; private set; }
     public DateTime IssueDate { get; private set; }
-    public decimal SubtotalService { get; private set; }
-    public decimal SubtotalMileageAllowance { get; private set; }
-    public decimal Total { get; private set; }
+    public double SubtotalService { get; private set; }
+    public double SubtotalMileageAllowance { get; private set; }
+    public double Total { get; private set; }
 
 
-    public Invoice(int id, string? description, DateTime issueDate, decimal subtotalService, decimal subtotalMileageAllowance, decimal total)
+    public Invoice(long id, string? description, DateTime issueDate, double subtotalService, double subtotalMileageAllowance, double total)
     {
         Id = id;
         Description = description;
